add items to cart

+++ server/db/models/cartLineItem.js
@@ -17,8 +17,8 @@ const CartLineItem = db.define("cartLineItem", {
     type: Sequelize.INTEGER,
     allowNull: false
   },
-  unitPriceAtPurchase: Sequelize.INTEGER,
-  totalPriceAtPurchase: Sequelize.INTEGER
+  unitPriceAtPurchase: Sequelize.FLOAT,
+  totalPriceAtPurchase: Sequelize.FLOAT
 });
 
 module.exports = CartLineItem;
